Revert "增加了文章的html"
This reverts commit 4ec4fbe946333fe7dd84253c75520f59e3b120c5.

+++ src/main/java/me/qyh/blog/service/impl/FileServiceImpl.java
@@ -215,10 +215,12 @@ public class FileServiceImpl implements FileService {
 	@Override
 	public void clearDeletedCommonFile() {
 		List<CommonFile> toDeletes = commonFileDao.selectDeleted();
-		for (CommonFile cf : toDeletes) {
-			FileStore fs = getFileStore(cf);
-			if (fs.delete(cf)) {
-				commonFileDao.deleteById(cf.getId());
+		if (!toDeletes.isEmpty()) {
+			for (CommonFile cf : toDeletes) {
+				FileStore fs = getFileStore(cf);
+				if (fs.delete(cf)) {
+					commonFileDao.deleteById(cf.getId());
+				}
 			}
 		}
 	}
